pyproject.toml
[project]
name = "catalogo-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
  "django>=5.0,<5.1",
  "django-ninja>=1.3.0",
  "django-environ>=0.11.2",
  "uvicorn>=0.34.0",
  "django-ninja-jwt>=5.3.5",
  "django-ninja-extra>=0.22.0",
  "pillow>=11.1.0",
  "django-libsql>=0.1.3",
  "django-cors-headers>=4.6.0",
]

[project.optional-dependencies]
ci = [
  "black>=24.10.0",
  "ruff>=0.6.9",
]
tests = [
  "ipython>=8.31.0",
  "pytest>=8.3.4",
  "pytest-cov>=6.0.0",
]

[tool.uv]
dev-dependencies = [
  "black>=24.10.0",
  "ruff>=0.6.9",
  "django-extensions>=3.2.3",
]

[tool.ruff]
line-length = 120
lint.select = ["I00", "E", "F", "W"]

[tool.black]
line-length = 120
target-version = ['py39']

# [tool.uv.dependencies]
# django = "^5.1.2"
# django-ninja = "^1.3.0"
# httpx = "^0.27.0"
# loguru = "^0.7.2"
# python = "^3.12"
# uvicorn = "^0.29.0"
